requirements.txt
streamlit
google-generativeai
pinecone
langgraph
langchain-google-genai
langchain-community
pymongo
python-dotenv
pypdf
pillow
tiktoken
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
bcrypt
googletrans==4.0.0rc1
requests
gTTS
SpeechRecognition
pyaudio
